pyproject.toml
[project]
name = "docker-gam"
version = "0.1.0"
description = "A metadata-driven management tool for organizing and controlling multiple Docker Compose stacks."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Jesse Johnson", email = "oss@fictiontoreality.dev"}
]
keywords = ["docker", "docker-compose", "container", "management", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
gam = "gam.cli:main"

[project.urls]
Homepage = "https://github.com/fictiontoreality/gam"
Repository = "https://github.com/fictiontoreality/gam"
Issues = "https://github.com/fictiontoreality/gam/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gam"]
